♻️refactor : 게시글 작성 이미지 적용
